refactor: rename core/transcode directory to core/stream
