+++ core/stream/limitations.go
@@ -1,4 +1,4 @@
-package transcode
+package stream
 
 import (
 	"strconv"
