+++ core/stream/media_streamer.go
@@ -1,4 +1,4 @@
-package transcode
+package stream
 
 import (
 	"context"
