+++ core/stream/token_test.go
@@ -1,4 +1,4 @@
-package transcode
+package stream
 
 import (
 	"context"
